backend/ml-service/requirements.txt
kafka-python
numpy
pandas
scikit-learn
tensorflow
